im to stupid to fix it

+++ hurt/util/array.d
@@ -4,29 +4,29 @@ import hurt.conv.conv;
 import hurt.exception.nullexception;
 import hurt.exception.outofrangeexception;
 
-pure void arrayCopy(T)(T[] src, in uint sOffset, T[] drain, in uint dOffset, 
-		in uint number) {
+pure void arrayCopy(T)(T[] src, in size_t sOffset, T[] drain, in size_t dOffset, 
+		in size_t number) {
 	if(src is null)
 		throw new NullException("Source is null");
 	if(drain is null)
 		throw new NullException("Drain is null");
 	if(sOffset + number > src.length) {
 		throw new OutOfRangeException("With this offset " 
-			~ conv!(uint, string)(sOffset) ~ " and this number "
-			~ conv!(uint, string)(number) ~ 
+			~ conv!(size_t, string)(sOffset) ~ " and this number "
+			~ conv!(size_t, string)(number) ~ 
 				" and Out of Bound Error will occur " 
 			~ "because the src array is to short. The array length is " 
-			~ conv!(uint,string)(src.length));
+			~ conv!(size_t,string)(src.length));
 	} else  if(dOffset + number > drain.length) {
 		throw new OutOfRangeException("With this offset " 
-			~ conv!(uint, string)(dOffset) ~ " and this number "
-			~ conv!(uint, string)(number) ~ 
+			~ conv!(size_t, string)(dOffset) ~ " and this number "
+			~ conv!(size_t, string)(number) ~ 
 				" and Out of Bound Error will occur " 
 			~ "because the drain array is to short. The array length is " 
-			~ conv!(uint,string)(drain.length));
+			~ conv!(size_t,string)(drain.length));
 	}
 	
-	for(uint idx = 0; idx < number; idx++) {
+	for(size_t idx = 0; idx < number; idx++) {
 		drain[dOffset + idx] = src[sOffset + idx];
 	}
 }
